wifi: Use int64_t for frequencies

+++ src/wifi/model/he/he-phy.cc
@@ -1229,9 +1229,9 @@ HePhy::GetPer20MHzDurations(const Ptr<const WifiPpdu> ppdu)
 
         if (ppdu)
         {
-            const uint16_t subchannelMinFreq =
+            const int64_t subchannelMinFreq =
                 m_wifiPhy->GetFrequency() - (m_wifiPhy->GetChannelWidth() / 2) + (index * 20);
-            const uint16_t subchannelMaxFreq = subchannelMinFreq + 20;
+            const int64_t subchannelMaxFreq = subchannelMinFreq + 20;
             const auto ppduBw = ppdu->GetTxVector().GetChannelWidth();
 
             if (ppduBw <= m_wifiPhy->GetChannelWidth() &&
@@ -1350,7 +1350,7 @@ HePhy::GetTxPowerSpectralDensity(double txPowerW,
     const auto& txVector = ppdu->GetTxVector();
     const auto& centerFrequencies = ppdu->GetTxCenterFreqs();
     auto channelWidth = txVector.GetChannelWidth();
-    auto printFrequencies = [](const std::vector<uint16_t>& v) {
+    auto printFrequencies = [](const std::vector<int64_t>& v) {
         std::stringstream ss;
         for (const auto& centerFrequency : v)
         {
@@ -1442,7 +1442,7 @@ HePhy::GetTxPowerSpectralDensity(double txPowerW,
     }
 }
 
-std::vector<uint16_t>
+std::vector<int64_t>
 HePhy::GetCenterFrequenciesForNonHePart(Ptr<const WifiPpdu> ppdu, uint16_t staId) const
 {
     NS_LOG_FUNCTION(this << ppdu << staId);
@@ -1459,7 +1459,7 @@ HePhy::GetCenterFrequenciesForNonHePart(Ptr<const WifiPpdu> ppdu, uint16_t staId
         HeRu::RuSpec nonOfdmaRu =
             HeRu::FindOverlappingRu(currentWidth, ru, HeRu::GetRuType(nonOfdmaWidth));
 
-        uint16_t startingFrequency = centerFrequencies.front() - (currentWidth / 2);
+        int64_t startingFrequency = centerFrequencies.front() - (currentWidth / 2);
         centerFrequencies.front() =
             startingFrequency +
             nonOfdmaWidth * (nonOfdmaRu.GetPhyIndex(
@@ -1858,7 +1858,7 @@ HePhy::GetRxPpduFromTxPpdu(Ptr<const WifiPpdu> ppdu)
 std::vector<WifiSpectrumBandIndices>
 HePhy::ConvertHeRuSubcarriers(ChannelWidthMhz bandWidth,
                               ChannelWidthMhz guardBandwidth,
-                              const std::vector<uint16_t>& centerFrequencies,
+                              const std::vector<int64_t>& centerFrequencies,
                               ChannelWidthMhz totalWidth,
                               uint32_t subcarrierSpacing,
                               HeRu::SubcarrierRange subcarrierRange,
