+++ src/wifi/test/wifi-phy-thresholds-test.cc
@@ -40,7 +40,7 @@ using namespace ns3;
 NS_LOG_COMPONENT_DEFINE("WifiPhyThresholdsTest");
 
 static const uint8_t CHANNEL_NUMBER = 36;
-static const uint32_t FREQUENCY = 5180;   // MHz
+static const int64_t FREQUENCY = 5180;    // MHz
 static const uint16_t CHANNEL_WIDTH = 20; // MHz
 
 /**
